Вынес вспомогательные методы для тестов в отдельный файл
